♻️ refactor: review 도메인 생성 및 전시, 공간 review 해당 도메인으로 위치 변경 및 전시 리뷰 생성
