+++ src/main/java/com/project/team5backend/domain/image/entity/ExhibitionReviewImage.java
@@ -1,11 +1,9 @@
 package com.project.team5backend.domain.image.entity;
 
-import com.project.team5backend.domain.exhibition.review.entity.ExhibitionReview;
+import com.project.team5backend.domain.review.exhibition.entity.ExhibitionReview;
 import com.project.team5backend.global.entity.BaseCreateDeleteEntity;
 import jakarta.persistence.*;
 import lombok.*;
-
-import java.time.LocalDateTime;
 
 @Entity
 @Getter
