+++ src/main/java/com/project/team5backend/domain/review/exhibition/entity/ExhibitionReview.java
@@ -1,0 +1,46 @@
+package com.project.team5backend.domain.review.exhibition.entity;
+
+import com.project.team5backend.domain.exhibition.entity.Exhibition;
+import com.project.team5backend.domain.image.entity.ExhibitionReviewImage;
+import com.project.team5backend.domain.user.entity.User;
+import com.project.team5backend.global.entity.BaseTimeEntity;
+import jakarta.persistence.*;
+import lombok.*;
+import org.hibernate.annotations.BatchSize;
+
+import java.util.List;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class ExhibitionReview extends BaseTimeEntity {
+
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "exhibition_review_id")
+    private Long id;
+
+    private String content;
+
+    private Double rating;
+
+    @Column(name = "is_deleted")
+    private boolean isDeleted;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "exhibition_id")
+    private Exhibition exhibition;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    @OneToMany(mappedBy = "exhibitionReview", cascade = CascadeType.ALL, orphanRemoval = true)
+    @BatchSize(size = 100)
+    private List<ExhibitionReviewImage> exhibitionReviewImages;
+
+    public void softDelete() {
+        isDeleted = true;
+    }
+}
