+++ src/main/java/com/project/team5backend/domain/review/exhibition/service/command/ExhibitionReviewCommandServiceImpl.java
@@ -1,0 +1,105 @@
+package com.project.team5backend.domain.review.exhibition.service.command;
+
+import com.project.team5backend.domain.review.exhibition.dto.response.ExhibitionReviewResDTO;
+import com.project.team5backend.domain.exhibition.entity.Exhibition;
+import com.project.team5backend.domain.user.exception.UserErrorCode;
+import com.project.team5backend.domain.user.exception.UserException;
+import com.project.team5backend.global.entity.enums.Status;
+import com.project.team5backend.domain.exhibition.exception.ExhibitionErrorCode;
+import com.project.team5backend.domain.exhibition.exception.ExhibitionException;
+import com.project.team5backend.domain.exhibition.repository.ExhibitionRepository;
+import com.project.team5backend.domain.review.exhibition.converter.ExhibitionReviewConverter;
+import com.project.team5backend.domain.review.exhibition.dto.request.ExhibitionReviewReqDTO;
+import com.project.team5backend.domain.review.exhibition.entity.ExhibitionReview;
+import com.project.team5backend.domain.review.exhibition.exception.ExhibitionReviewErrorCode;
+import com.project.team5backend.domain.review.exhibition.exception.ExhibitionReviewException;
+import com.project.team5backend.domain.review.exhibition.repository.ExhibitionReviewRepository;
+import com.project.team5backend.domain.image.converter.ImageConverter;
+import com.project.team5backend.domain.image.entity.ExhibitionReviewImage;
+import com.project.team5backend.domain.image.exception.ImageErrorCode;
+import com.project.team5backend.domain.image.exception.ImageException;
+import com.project.team5backend.domain.image.repository.ExhibitionReviewImageRepository;
+import com.project.team5backend.domain.image.service.command.ImageCommandService;
+import com.project.team5backend.domain.user.entity.User;
+import com.project.team5backend.domain.user.repository.UserRepository;
+import com.project.team5backend.global.util.S3Uploader;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.time.LocalDate;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+@Transactional
+public class ExhibitionReviewCommandServiceImpl implements ExhibitionReviewCommandService {
+
+    private final UserRepository userRepository;
+    private final ExhibitionRepository exhibitionRepository;
+    private final ExhibitionReviewRepository exhibitionReviewRepository;
+    private final ExhibitionReviewImageRepository exhibitionReviewImageRepository;
+    private final ImageCommandService imageCommandService;
+    private final S3Uploader s3Uploader;
+    @Override
+    public ExhibitionReviewResDTO.ExReviewCreateResDTO createExhibitionReview(Long exhibitionId, Long userId, ExhibitionReviewReqDTO.createExReviewReqDTO createExhibitionReviewReqDTO, List<MultipartFile> images) {
+        LocalDate current = LocalDate.now();
+
+        Exhibition exhibition = exhibitionRepository.findByIdAndIsDeletedFalseAndStatusApproveAndOpening(exhibitionId, current, Status.APPROVED)
+                .orElseThrow(()-> new ExhibitionException(ExhibitionErrorCode.EXHIBITION_NOT_FOUND));
+        User user = userRepository.findByIdAndIsDeletedFalse(userId)
+                .orElseThrow(()-> new UserException(UserErrorCode.USER_NOT_FOUND));
+
+        ExhibitionReview exhibitionReview = ExhibitionReviewConverter.toEntity(createExhibitionReviewReqDTO, exhibition, user);
+        exhibitionReviewRepository.save(exhibitionReview);
+
+        List<ExhibitionReviewImage> reviewImages = images.stream()
+                .map(file -> {
+                    String url = s3Uploader.upload(file, "exhibitionReviews");
+                    return ImageConverter.toExhibitionReviewImage(exhibitionReview, url);
+                })
+                .toList();
+        exhibitionReviewImageRepository.saveAll(reviewImages);
+
+        // 5) 리뷰 평균/카운트 갱신
+        double rating = exhibitionReview.getRating();
+        exhibitionRepository.applyReviewCreated(exhibitionId, rating);
+
+        return ExhibitionReviewConverter.toExReviewCreateResDTO(exhibitionReview.getId());
+    }
+    @Override
+    public void deleteExhibitionReview(Long exhibitionReviewId, String email) {
+        ExhibitionReview exhibitionReview = exhibitionReviewRepository.findByIdAndIsDeletedFalse(exhibitionReviewId)
+                .orElseThrow(() -> new ExhibitionReviewException(ExhibitionReviewErrorCode.EXHIBITION_REVIEW_NOT_FOUND));
+
+        if (exhibitionReview.isDeleted()) return;
+
+        if (!Objects.equals(exhibitionReview.getUser().getEmail(), email)) {
+            throw new ExhibitionReviewException(ExhibitionReviewErrorCode.EXHIBITION_REVIEW_FORBIDDEN);
+        }
+
+        exhibitionReview.softDelete();
+
+        // 전시이미지 소프트 삭제
+        List<ExhibitionReviewImage> images = exhibitionReviewImageRepository.findByExhibitionReviewId(exhibitionReviewId);
+        images.forEach(ExhibitionReviewImage::deleteImage);
+        List<String> keys = images.stream().map(ExhibitionReviewImage::getImageUrl).toList();
+
+        // 리뷰 평균/카운트 갱신
+        double rating = exhibitionReview.getRating();
+        exhibitionRepository.applyReviewDeleted(exhibitionReview.getExhibition().getId(), rating);
+
+        // s3 보존 휴지통 prefix로 이동시키기
+        try{
+            imageCommandService.moveToTrashPrefix(keys);
+        } catch (ImageException e) {
+            throw new ImageException(ImageErrorCode.S3_MOVE_TRASH_FAIL);
+        }
+    }
+
+}
