+++ src/main/java/com/project/team5backend/domain/review/space/dto/request/SpaceReviewReqDTO.java
@@ -1,0 +1,9 @@
+package com.project.team5backend.domain.review.space.dto.request;
+
+public class SpaceReviewReqDTO {
+
+    public record CreateSpaceReviewReqDTO(
+            Double rating,
+            String content
+    ){}
+}
