+++ src/main/java/com/project/team5backend/domain/review/space/exception/SpaceReviewException.java
@@ -1,0 +1,9 @@
+package com.project.team5backend.domain.review.space.exception;
+
+import com.project.team5backend.global.apiPayload.exception.CustomException;
+
+public class SpaceReviewException extends CustomException {
+  public SpaceReviewException(SpaceReviewErrorCode errorCode) {
+    super(errorCode);
+  }
+}
