+++ src/main/java/com/project/team5backend/domain/review/space/service/query/SpaceReviewQueryService.java
@@ -1,0 +1,11 @@
+package com.project.team5backend.domain.review.space.service.query;
+
+
+import com.project.team5backend.domain.review.space.dto.response.SpaceReviewResDTO;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+
+public interface SpaceReviewQueryService {
+    Page<SpaceReviewResDTO.DetailSpaceReviewResDTO> getSpaceReviewList(long spaceId, Pageable pageable);
+    SpaceReviewResDTO.DetailSpaceReviewResDTO getSpaceReviewDetail(long spaceReviewId);
+}
