--- src/main/java/com/project/team5backend/domain/space/review/dto/request/SpaceReviewReqDTO.java
@@ -1,9 +1,0 @@
-package com.project.team5backend.domain.space.review.dto.request;
-
-public class SpaceReviewReqDTO {
-
-    public record CreateSpaceReviewReqDTO(
-            Double rating,
-            String content
-    ){}
-}
